+++ lib/capypage/element.rb
@@ -3,25 +3,24 @@ require 'active_support/core_ext/object/blank'
 
 module Capypage
   class Element
-    attr_accessor :selector, :prefix, :finder_options, :base_element
+    attr_accessor :selector, :finder_options, :base_element
 
-    def initialize(selector, prefix = nil, options = {}, &block)
+    def initialize(selector, options = {}, &block)
       options.reverse_merge! :match => :first
       @finder_options = options.clone
       @selector = selector
-      @prefix = prefix
-      @base_element = finder_options[:base_element] || capybara_page
+      @base_element = finder_options[:base_element]
       block.call(self) if block.present?
     end
 
     def element(name, selector, options = {})
       base = self
-      define_singleton_method(name) { Element.new(selector, prefix, options.reverse_merge!(:match => :first, :base_element => base)) }
+      define_singleton_method(name) { Element.new(selector, options.reverse_merge!(:match => :first, :base_element => base)) }
     end
 
     def elements(name, selector, options = {}, &block)
       base = self
-      define_singleton_method(name) { Elements.new(selector, prefix, options.reverse_merge!(:base_element => base), &block) }
+      define_singleton_method(name) { Elements.new(selector, options.reverse_merge!(:base_element => base), &block) }
     end
 
     def present?
@@ -40,11 +39,7 @@ module Capypage
     end
 
     def element_selector
-      [prefix, selector].compact.join(" ")
-    end
-
-    def capybara_page
-      Capybara.current_session
+      selector
     end
 
     def capybara_finder_options(options = {})