+++ lib/capypage/elements.rb
@@ -6,19 +6,19 @@ module Capypage
     delegate :each, :size, :[],
              :to => :all
 
-    def initialize(parent_selector, children_selector, prefix = nil, options = {}, &block)
-      parent_selector_options = options.merge :base_element => Element.new(parent_selector, prefix, options)
+    def initialize(parent_selector, children_selector, options = {}, &block)
+      parent_selector_options = options.merge :base_element => Element.new(parent_selector, options)
 
-      super(children_selector, prefix, parent_selector_options)
+      super(children_selector, parent_selector_options)
       @child_dsl_block = block
     end
 
     def find_by_text(text, options = {})
-      Element.new(selector, prefix, finder_options.merge(options).merge(:text => text), &child_dsl_block)
+      Element.new(selector, finder_options.merge(options).merge(:text => text), &child_dsl_block)
     end
 
     def find_by_index(index)
-      Element.new("#{element_selector}:nth-child(#{index + 1})", prefix, finder_options, &child_dsl_block)
+      Element.new("#{element_selector}:nth-child(#{index + 1})", finder_options, &child_dsl_block)
     end
 
     private