+++ lib/capypage/page.rb
@@ -8,8 +8,7 @@ module Capypage
     attr_accessor :prefix, :finder_options
 
     def initialize(prefix = nil, options = {})
-      @prefix = prefix
-      @finder_options = options
+      @finder_options = options.merge(:base_element => prefix ? Element.new(prefix, :base_element => Capybara.current_session) : Capybara.current_session)
     end
 
     class <<self
@@ -18,18 +17,15 @@ module Capypage
       end
 
       def element(name, selector, options = {})
-        options.reverse_merge! :match => :first
-        define_method(name) { Element.new(selector, prefix, options) }
+        define_method(name) { Element.new(selector, finder_options.reverse_merge(options)) }
       end
 
       def elements(name, parent_selector, children_selector, options = {}, &block)
-        define_method(name) { Elements.new(parent_selector, children_selector, prefix, options, &block) }
+        define_method(name) { Elements.new(parent_selector, children_selector, finder_options.reverse_merge(options), &block) }
       end
 
       def section(name, section, selector, options = {})
-        define_method name do
-          section.new(selector, options)
-        end
+        define_method(name) { section.new(selector, options) }
       end
     end