+++ lib/capypage/section.rb
@@ -1,5 +1,4 @@
 module Capypage
   class Section < Page
-    alias_method :selector, :prefix
   end
 end