+++ CHANGELOG.md
@@ -1,6 +1,13 @@
+v 0.2.1
+
+* Only page takes a prefix now
+* Cleaned up section implementation
+
+
 v 0.2.0
 
 * Not backward compatible - breaks the elements api in page to have better assertions on the elements level
+
 
 v 0.1.1